add grid

+++ client/src/pages/DashboardPage.tsx
@@ -1,4 +1,6 @@
 import { useCallback, useEffect, useState } from "react";
+import GridLayout from "react-grid-layout";
+import "react-grid-layout/css/styles.css";
 import { useParams } from "react-router-dom";
 import { ChatDto } from "../../../server/src/db/models/chats";
 import { WsOutputMessage } from "../../../shared/types";
@@ -14,6 +16,8 @@ function DashboardPage() {
   const params = useParams<{ id: string }>();
   const { lastMessage, deleteChat } = useWs();
 
+  const [showChatStarter, setShowChatStarter] = useState(false);
+
   const handleWsMessage = useCallback((msg: WsOutputMessage) => {
     if (msg.type === "CHAT_STARTED") {
       setChats((prev) => [...prev, msg.payload]);
@@ -26,6 +30,10 @@ function DashboardPage() {
 
   const handleDeleteChat = useCallback(
     (id: string) => {
+      if (!confirm("Are you sure you want to delete this chat?")) {
+        return;
+      }
+
       deleteChat(id);
       setChats((prev) => prev.filter((chat) => chat.uuid !== id));
     },
@@ -57,6 +65,11 @@ function DashboardPage() {
       // const db = await api.get<DashboardDto[]>(`/dashboards/${params.id}`);
       const chts = await api.get<ChatDto[]>(`/dashboards/${params.id}/chats`);
       setChats(chts);
+
+      if (chts.length === 0) {
+        console.log("No chats, showing chat starter");
+        setShowChatStarter(true);
+      }
     }
 
     fetchData();
@@ -66,17 +79,69 @@ function DashboardPage() {
     return null;
   }
 
+  const layout: GridLayout.Layout[] = chats.map((chat, i) => ({
+    i: chat.uuid,
+    x: i * 4,
+    y: i * 2,
+    w: 4,
+    h: 16,
+  }));
+
   return (
     <main className={`app ${chats.length === 0 ? "no-chats" : ""}`}>
-      {chats.map((chat) => (
-        <Chat
-          key={chat.uuid}
-          id={chat.uuid}
-          name={chat.profileName}
-          onDelete={() => handleDeleteChat(chat.uuid)}
-        />
-      ))}
-      {profiles && <ChatStarter profiles={profiles} dbUuid={params.id} />}
+      <GridLayout
+        className="layout"
+        layout={layout}
+        cols={12}
+        rowHeight={30}
+        width={1200}
+        onLayoutChange={(...data) => {
+          console.log("onLayoutChange", data);
+        }}
+        draggableHandle=".drag-handle"
+      >
+        {chats.map((chat) => (
+          <div key={chat.uuid} className="chat-wrapper">
+            <div className="top-menu">
+              <div className="left">
+                <div className="profile-name">{chat.profileName}</div>
+              </div>
+              <div className="right">
+                <button
+                  className="closeChat"
+                  title="Delete chat"
+                  aria-label="Delete chat"
+                  aria-hidden="true"
+                  onClick={() => handleDeleteChat(chat.uuid)}
+                >
+                  x
+                </button>
+
+                <div className="drag-handle">::</div>
+              </div>
+            </div>
+            <div className="body">
+              <Chat id={chat.uuid} />
+            </div>
+          </div>
+        ))}
+      </GridLayout>
+      {profiles && (
+        <>
+          <div className="db-menu">
+            <button onClick={() => setShowChatStarter((prev) => !prev)}>
+              {showChatStarter ? "x" : "+"}
+            </button>
+          </div>
+          {showChatStarter && (
+            <ChatStarter
+              profiles={profiles}
+              dbUuid={params.id}
+              onStarted={() => setShowChatStarter(false)}
+            />
+          )}
+        </>
+      )}
     </main>
   );
 }
